feat: Car 객체의 String name 인스턴스 변수 객체로 포장
- Name 객체 생성
